+++ packages/purchasely/ios/PurchaselyRN.h
@@ -15,6 +15,8 @@
 
 @property (class, nonatomic, copy) RCTPromiseResolveBlock purchaseResolve;
 
+@property (class, nonatomic, strong) UIViewController *sharedViewController;
+
 @property (class, nonatomic, strong) NSMutableArray<id<PLYPresentation>> *presentationsLoaded;
 
 @property (nonatomic, assign) Boolean shouldReopenPaywall;