+++ packages/purchasely/src/startBuilder.ts
@@ -50,7 +50,7 @@ export class PurchaselyBuilder {
      *
      * @internal
      */
-    static bridgeVersion = '6.0.0-rc.2';
+    static bridgeVersion = '6.0.0-rc.3';
 
     private constructor(private readonly state: StartBuilderState) {}